+++ packages/select/src/FilterMultiSelect/provider/SelectionProvider/SelectionProvider.tsx
@@ -4,7 +4,7 @@ import { Item } from "@react-stately/collections"
 import { ListState, useListState } from "@react-stately/list"
 import { SelectionMode, Node, Selection } from "@react-types/shared"
 import { VisuallyHidden } from "@kaizen/a11y"
-import { ItemType } from "../../types"
+import { ItemType } from "../../../types"
 
 export interface SelectionProviderProps {
   selectionMode: SelectionMode